add github action and fixes
