--- .env.example
@@ -1,11 +1,0 @@
-# Link Plugin Environment Configuration
-
-# Development Settings
-VITE_HOST=localhost
-VITE_PORT=3005
-
-# Production Plugin URL (used when VITE_DEV is false)
-# VITE_PLUGIN_URL=https://your-domain.com/plugins/link-plugin/plugin.js
-
-# Build Configuration
-# VITE_PLUGIN_URL=/plugins/link-plugin/plugin.js  # Relative path for production
