+++ dev/main.ts
@@ -2,13 +2,18 @@ import { createApp } from 'vue';
 import './style.css';
 import App from './App.vue';
 import { createRouter, createWebHistory } from 'vue-router';
+import PluginEnvironment from './components/PluginEnvironment.vue';
 
 const router = createRouter({
   history: createWebHistory(),
   routes: [
     {
+      path: '/',
+      component: PluginEnvironment,
+    },
+    {
       path: '/:id',
-      component: App,
+      component: PluginEnvironment,
     }
   ]
 });
